Move tests for soft_cross_entropy out of label smoothing tests
