+++ tests/models/test_loss.py
@@ -5,7 +5,7 @@ import pytest
 import torch
 import torch.nn.functional as F
 
-from composer.models.loss import MIoU, ensure_targets_one_hot, soft_cross_entropy
+from composer.models.loss import MIoU, ensure_targets_one_hot, soft_cross_entropy, _infer_target_type
 
 
 @pytest.fixture
@@ -18,10 +18,7 @@ def block_2D_targets():
     return targets
 
 
-# (input shape)
-@pytest.fixture(params=[(3, 5), (7, 11, 13, 17)])
-def fake_input_target_pairs(request):
-    input_shape = request.param
+def fake_input_target_pairs(input_shape):
     num_classes = input_shape[1]
     reduced_input_shape = list(input_shape)
     reduced_input_shape.pop(1)
@@ -35,6 +32,20 @@ def fake_input_target_pairs(request):
             i_expanded.insert(1, value)
             targets_one_hot[tuple(i_expanded)] = 1.0
     return input, targets_idx, targets_one_hot
+
+
+def generate_tensors():
+    return [
+        # Binary classification
+        fake_input_target_pairs((64, 2)),
+        # Classification
+        fake_input_target_pairs((64, 10))
+    ]
+
+
+def xfail(val):
+    """shorthand to mark xfail parameters."""
+    return pytest.param(val, marks=pytest.mark.xfail)
 
 
 def test_miou(block_2D_targets):
@@ -72,14 +83,43 @@ def test_miou(block_2D_targets):
     assert torch.isclose(miou.compute(), torch.tensor(100 / 7, dtype=torch.double))
 
 
-def test_ensure_targets_one_hot(fake_input_target_pairs):
-    input, targets_idx, targets_one_hot = fake_input_target_pairs
+@pytest.mark.parametrize('tensors', generate_tensors())
+def test_ensure_targets_one_hot(tensors):
+    input, targets_idx, targets_one_hot = tensors
     targets_one_hot_test = ensure_targets_one_hot(input, targets_idx)
     torch.testing.assert_allclose(targets_one_hot, targets_one_hot_test)
 
 
-def test_soft_cross_entropy(fake_input_target_pairs):
-    input, targets_idx, targets_one_hot = fake_input_target_pairs
+@pytest.mark.parametrize('tensors', generate_tensors())
+def test_soft_cross_entropy(tensors):
+    input, targets_idx, targets_one_hot = tensors
     loss_idx = F.cross_entropy(input, targets_idx, ignore_index=-1)
     loss_soft = soft_cross_entropy(input, targets_one_hot)
     torch.testing.assert_allclose(loss_idx, loss_soft)
+
+
+@pytest.mark.parametrize('tensors', generate_tensors())
+class TestSoftCrossEntropy:
+
+    def test_infer_target_type(self, tensors):
+        input, target_indices, target_onehot = tensors
+        assert _infer_target_type(input, target_indices) == 'indices'
+        assert _infer_target_type(input, target_onehot) == 'one_hot'
+
+    @pytest.mark.parametrize('reduction', ['mean', 'sum'])
+    @pytest.mark.parametrize('use_weights', [xfail(True), False])
+    def test_soft_cross_entropy(self, tensors, use_weights, reduction):
+        (input, target_indices, target_onehot) = tensors
+        if use_weights:
+            num_classes = target_onehot.shape[1]
+            weights = torch.rand(size=[num_classes])
+        else:
+            weights = None
+
+        loss_args = dict(ignore_index=-1, weight=weights, reduction=reduction)
+        loss_indices = soft_cross_entropy(input, target_indices, **loss_args)
+        loss_onehot = soft_cross_entropy(input, target_onehot, **loss_args)
+        loss_reference = F.cross_entropy(input, target_indices, **loss_args)
+
+        torch.testing.assert_allclose(loss_indices, loss_onehot)
+        torch.testing.assert_allclose(loss_indices, loss_reference)
